+++ packages/orq-rc/src/funcs/knowledgeList.ts
@@ -165,7 +165,7 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, components.KnowledgeBasesServiceListResponse$inboundSchema),
-    M.fail("4XX"),
+    M.fail([401, "4XX"]),
     M.fail("5XX"),
   )(response, req);
   if (!result.ok) {