+++ packages/orq-rc/src/funcs/modelsValidate.ts
@@ -152,7 +152,7 @@ async function $do(
     | SDKValidationError
   >(
     M.nil(200, z.void()),
-    M.fail([400, 415, "4XX"]),
+    M.fail([400, 401, 415, "4XX"]),
     M.fail("5XX"),
   )(response, req);
   if (!result.ok) {