+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-14T10:10:20.455Z",
+    "2026-09-14T11:45:22.728Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {