+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2FPB4WDXGAPNF4DYA1DJPX8"),
+  _id: z.string().default("tool_01M2FVS3BWY3NFRWSBQT1TAG8P"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2FPB4WC5ZC733BX2SMT02ZF"),
+  _id: z.string().default("tool_01M2FVS3BVVW6044C1YXZMQJEE"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2FPB4WCWREQX4NAQECSRF9Z"),
+  _id: z.string().default("tool_01M2FVS3BV9YRK875QDWP6RCVZ"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2FPB4WBF1BNZ6SX8263G8Y5"),
+  _id: z.string().default("tool_01M2FVS3BVT6S8JKZ33SHECAV9"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),