+++ packages/orq-rc/RELEASES.md
@@ -12329,3 +12329,13 @@ Based on:
 - [typescript v4.15.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.44 - packages/orq-rc
+
+## 2026-09-14 11:48:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.45 - packages/orq-rc